requirements.txt
PyYAML
argparse
pandas
yfinance
matplotlib
requests
backtrader
